让g++开心。 加了一点功能。

+++ bis/Data/Referrence.cpp
@@ -8,7 +8,7 @@ void mule::Data::Referrence::Read(xybase::Stream *stream, DataHandler *dataHandl
 	size_t loc = stream->Tell();
 	stream->Seek(ptr, 0);
 	referent->Read(stream, dataHandler);
-	dataHandler->AppendMetadatum("addr", (unsigned long long)ptr);
+	dataHandler->AppendMetadatum("addr", (uint64_t)ptr);
 	stream->Seek(loc, 0);
 }
